backend/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# 数据验证
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# JWT和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2

# Redis
redis==5.0.1

# 邮件
aiosmtplib==3.0.1

# HTTP客户端
httpx==0.25.1

# 工具库
pyyaml==6.0.1

